@@ -3,22 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/jest-dom": "^5.15.1",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
+    "@types/react": "^17.0.37",
     "@types/react-dom": "^17.0.0",
     "@types/react-router-dom": "^5.3.2",
     "axios": "^0.24.0",
     "classnames": "^2.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "sass": "^1.43.4",
-    "typescript": "^4.1.2",
+    "react-router-dom": "^6.2.1",
+    "sass": "^1.44.0",
+    "typescript": "^4.5.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -47,6 +46,7 @@
     ]
   },
   "devDependencies": {
-    "prettier": "^2.4.1"
+    "prettier": "^2.5.0",
+    "react-scripts": "4.0.3"
   }
 }
