@@ -1,27 +1,20 @@
-import {
-  BrowserRouter as Router,
-  Redirect,
-  Route,
-  Switch,
-} from "react-router-dom";
+import { Navigate, Route, Routes } from "react-router-dom";
+import Home from "./pages/Home/Home";
+import NotFound from "./pages/NotFound/NotFound";
 import "./styles/main.scss";
-import HomeView from "./views/home/HomeView";
-import NotFoundView from "./views/notFound/NotFoundView";
 
 const App = () => {
   return (
-    <Router>
-      <Switch>
-        {/* Routes here */}
-        <Route exact path="/home" component={HomeView} />
+    <Routes>
+      {/* Routes here */}
+      <Route path="/home" element={<Home />} />
 
-        {/* Any redirects */}
-        <Redirect exact path="/" to="/home" />
+      {/* Any redirects */}
+      <Route path="/" element={<Navigate replace to="/home" />} />
 
-        {/* Last is a catch-all route that will show a not found view */}
-        <Route component={NotFoundView} />
-      </Switch>
-    </Router>
+      {/* Last is a catch-all route that will show a not found view */}
+      <Route path="*" element={<NotFound />} />
+    </Routes>
   );
 };
 
