@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import Layout from "../../components/layout/Layout";
+import Layout from "../../components/Layout/Layout";
 import { ErrorCode } from "../../enums/ErrorCode";
 
 const NotFoundView = () => (
